requirement.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.1
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.3.0
Django==4.2.10
django-admin-soft-dashboard==1.0.24
django-classy-tags==3.0.1
django-cms==3.8.2
django-filer==3.0.3
django-formtools==2.4
django-js-asset==2.0.0
django-mptt==0.16.0
django-phonenumber-field==8.0.0
django-polymorphic==3.1.0
django-sekizai==3.0.1
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==4.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.1
djangocms-googlemap==2.2.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.1.1
djangocms-picture==3.0.0
djangocms-style==3.1.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
easy-thumbnails==2.9
html5lib==1.1
httpie==3.2.4
idna==3.10
lxml==5.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
mysqlclient==2.2.7
phonenumberslite==8.13.55
pillow==11.1.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
PySocks==1.7.1
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
webencodings==0.5.1
